perf(config): Split TOKEN_ENCODINGS to avoid main-thread gpt-tokenizer load
`configSchema.ts` imported `TOKEN_ENCODINGS` from `TokenCounter.ts`, which
statically imports `gpt-tokenizer/GptEncoding` and
`gpt-tokenizer/resolveEncodingAsync`. Because `configSchema.ts` is on the
main-thread startup path (loaded by `defaultAction` → `configLoad`), every
CLI invocation paid the ~70ms cost of compiling the gpt-tokenizer ESM
module graph (`BytePairEncodingCore`, `constants`, `mapping`,
`functionCalling`, etc.) even though the actual tokenization runs only
inside the metrics worker pool.

This extracts `TOKEN_ENCODINGS` and `TokenEncoding` into a new file
`TokenEncoding.ts` that has no runtime dependencies beyond the string
literal tuple. `TokenCounter.ts` re-exports the constant and type so
external consumers of the public API (`src/index.ts`) keep working.
`configSchema.ts` now imports from the new file directly, bypassing the
gpt-tokenizer chain on the main thread.

Benchmark (30 alternating paired runs, `--quiet` on the repomix `src/`
tree, warm compile cache):

                 baseline   change   delta
  mean            1472 ms   1420 ms  -52 ms  (-3.5%)
  median          1470 ms   1412 ms  -58 ms  (-3.9%)
  trimmed mean    1468 ms   1409 ms  -59 ms  (-4.1%)

Paired comparison: change wins 21/30, ties 1/30, loses 8/30.

Isolated import timing (warm node cache):
  before: `import configSchema.js` = 98 ms
  after:  `import configSchema.js` = 21 ms

No functional change — `TOKEN_ENCODINGS` and `TokenEncoding` are still
re-exported from `TokenCounter.ts`, so both the internal worker path and
the public `index.ts` API surface are unchanged.
